requirements-generated.txt
# generated
coloredlogs
PyJWT
tartiflette>=1.0.0rc2
tartiflette-asgi
uvicorn
tartiflette_plugin_apollo_federation
motor
tartiflette_scalars>=0.0.6
aiohttp-cors
mongodb_streams>=0.0.7
